requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
black==23.3.0
boto3==1.34.152
botocore==1.34.152
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
cfgv==3.3.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.8
django-celery-beat==2.5.0
django-cors-headers==4.3.1
-e git+https://github.com/jazzband/django-debug-toolbar.git@199c2b3c6bc916d371bcfb11b0fcbbb690ee5e36#egg=django_debug_toolbar
django-extensions==3.2.3
django-filter==23.2
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-storages==1.14.4
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.12.0
flake8==6.0.0
fonttools==4.48.1
gunicorn==21.2.0
html5lib==1.1
identify==2.5.24
idna==3.6
inflection==0.5.1
interrogate==1.5.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
json-encoder==0.4.4
kombu==5.3.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.26.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.2
pandas==2.1.4
pathspec==0.11.1
phonenumbers==8.13.27
Pillow==10.1.0
platformdirs==3.5.0
pre-commit==3.3.1
prompt-toolkit==3.0.43
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.8.0
pyflakes==3.0.1
PyJWT==2.8.0
pyphen==0.14.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
s3transfer==0.10.2
salary==2.0
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
virtualenv==20.23.0
wcwidth==0.2.12
weasyprint==61.0
webencodings==0.5.1
whitenoise==6.6.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
zopfli==0.2.3
